Convert level compiler to f64 and fix BSP convexity termination
The level compiler now uses glam::DVec3 and f64 throughout for
geometric computation. Floating-point noise from interactive map
editors (TrenchBroom) was being amplified by cross-product plane
construction in f32, causing planes to be subtly tilted and tripping
downstream CSG/BSP epsilons. This matches the approach used by all
production Quake-family compilers (qbsp, ericw-tools).

The runtime engine and PRL file format remain f32 — only the in-memory
compile pipeline becomes f64. Precision conversions happen at exactly
five well-defined boundary sites: shambler_to_dvec3 at the parse input,
and dvec3_to_f32_array (or inline equivalents) at the four PRL encoder
outputs (portal vertices, BSP node planes, BSP leaf bounds, geometry
vertex buffer).

Also includes two compile-time bug fixes that are entangled with the
f64 conversion in the same files:

- BSP build_recursive now requires leaves to be both convex AND under
  the face count cap, instead of either/or. Non-convex face sets
  continue splitting until they become convex. Without this, leaves
  spanning both solid brush space and adjacent air gaps were
  classified solid by the centroid-based test, causing missing
  surfaces near floating brushes and broken portal generation around
  pillars.

- BSP score_plane and partition_faces distribute coplanar faces to
  the child they face toward (via normal dot product) rather than
  always sending them to the front child. This eliminates z-fighting
  and missing surfaces on parallel coplanar geometry.

New tests document the expected behavior for both compile-time fixes:

- room_with_pillar_produces_portals_on_both_sides
- floating_cube_air_space_leaves_stay_empty

+++ postretro-level-compiler/src/map_format.rs
@@ -41,7 +41,7 @@ impl MapFormat {
     /// normals are direction vectors; only the axis swizzle applies to them.
     ///
     /// For IdTech2: 1 Quake unit = 0.0254 m (exact, since 1 inch = 0.0254 m).
-    pub fn units_to_meters(&self) -> f32 {
+    pub fn units_to_meters(&self) -> f64 {
         match self {
             Self::IdTech2 => 0.0254,
             // IdTech3 and IdTech4 are unsupported — unreachable in practice
